Got the ortho viewer working more or less correctly.  Draws copies of elements from 3D scene without TEveProjectionManager.  Zoom needs to be improved by someone who understands TGLViewer and TGLCamera.
CopySceneMaker suggests restructure of EVE-based event display.  CopySceneMaker adds a scene that was created by ANOTHER MODULE to a viewer by getting that
scene by name from the TEveManager in the EveDisplay service.  This method uses communication between modules oustide of Principal objects (like art::Event)
which is NOT SUPPOSED TO BE USED.

If the technique in CopySceneMaker is permissable, then I propose to make a new module that will run all of the scene maker algorithms before the viewer makers.
We could then rewrite the viewer makers to only use CopySceneMaker-like algorithms or just always use them with CopySceneMaker.  Neither approach would change
how the EVE-based event display operates, but both seem to make it a lot neater.  I propose that this new module might be a Producer (if there can be producers that don't produce anything) to guarantee that it is always run before the viewer analyzers.